% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{anti_join.folded}
\alias{anti_join.folded}
\title{Anti-join Folded}
\usage{
\method{anti_join}{folded}(.data, ..., .dots)
}
\arguments{
\item{.data}{passed to next method}

\item{...}{passed to next method}

\item{.dots}{passed to next method}
}
\value{
folded
}
\description{
Anti-joins folded.
}
\keyword{internal}
