% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unfolded.R
\name{anti_join.unfolded}
\alias{anti_join.unfolded}
\title{Anti-join Unfolded}
\usage{
\method{anti_join}{unfolded}(.data, ..., .dots)
}
\arguments{
\item{.data}{passed to next method}

\item{...}{passed to next method}

\item{.dots}{passed to next method}
}
\value{
unfolded
}
\description{
Anti-joins unfolded.
}
\keyword{internal}
