% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/origami_SuperLearner.R
\name{aorder}
\alias{aorder}
\title{sort n-dimensional array (for multinomial SL support)}
\usage{
aorder(mat, index, along = 1)
}
\description{
sort n-dimensional array (for multinomial SL support)
}

