% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unfolded.R
\name{arrange.unfolded}
\alias{arrange.unfolded}
\title{Arrange Unfolded}
\usage{
\method{arrange}{unfolded}(.data, ...)
}
\arguments{
\item{.data}{passed to next method}

\item{...}{passed to next method}

\item{.dots}{passed to next method}
}
\value{
unfolded
}
\description{
Arranges unfolded.
}
\keyword{internal}
