% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{arrange_.folded}
\alias{arrange_.folded}
\title{Arrange Folded}
\usage{
\method{arrange_}{folded}(.data, ..., .dots)
}
\arguments{
\item{.data}{passed to next method}

\item{...}{passed to next method}

\item{.dots}{passed to next method}
}
\value{
folded
}
\description{
Arranges folded.
}
\keyword{internal}
