% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{as.csv.folded}
\alias{as.csv.folded}
\title{Save Folded Format as CSV}
\usage{
\method{as.csv}{folded}(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\value{
invisible folded (x)
}
\description{
Saves folded format as CSV.
}
\details{
Simply calls as.csv.data.frame.
}
\seealso{
\code{\link{fold.data.frame}} \code{\link{as.folded.character}}
}
