% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{as.folded.character}
\alias{as.folded.character}
\title{Coerce to Folded Format from Filepath}
\usage{
\method{as.folded}{character}(x, ...)
}
\arguments{
\item{x}{length-one character}

\item{...}{passed arguments}
}
\value{
folded
}
\description{
Coerces to folded format from character, treating \code{x} as a filepath.
}
