% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{as.folded.data.frame}
\alias{as.folded.data.frame}
\title{Coerce to Folded from Data Frame}
\usage{
\method{as.folded}{data.frame}(x, sort = TRUE, ...)
}
\arguments{
\item{x}{object}

\item{sort}{Should the result be sorted?}

\item{...}{passed arguments}
}
\value{
folded
}
\description{
Coerces to folded from data.frame.
}
\details{
Expects columns VARIABLE, META, and VALUE. Remaining columns are for classification and may be NA. Coerces VALUE to character. Removes duplicate records with warning. Sorts on non-value columns by default. Groups by columns other than VALUE. Assigns class \code{folded}.
}
