% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{as.folded.folded}
\alias{as.folded.folded}
\title{Coerce to Folded from Folded}
\usage{
\method{as.folded}{folded}(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\value{
folded data.frame
}
\description{
Coerces to folded from folded.  A non-operation.
}
\seealso{
\code{\link{fold.data.frame}} \code{\link{as.folded}}
}
\keyword{internal}
