% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{as.folded}
\alias{as.folded}
\title{Coerce to Folded Format}
\usage{
as.folded(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\description{
Coerces to folded format.
}
\seealso{
\code{\link{as.folded.character}} \code{\link{as.folded.folded}} \code{\link{as.folded.data.frame}}
}
\keyword{internal}
