% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaplot.R
\name{boxplot.folded}
\alias{boxplot.folded}
\title{Boxplot Method for Folded}
\usage{
\method{boxplot}{folded}(x, ...)
}
\arguments{
\item{x}{folded}

\item{...}{passed to \code{\link{boxplot.data.frame}}}
}
\description{
Boxplot for folded. Converts to data.frame with defined column attributes and calls data.frame method.
}
\examples{
data(eventsf)
boxplot(eventsf, SEX, WT, ref = 68)
}
\seealso{
Other bivariate plots: \code{\link{metaplot.folded}},
  \code{\link{scatter.folded}}
}
