% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fold_funs.R
\name{check_id_and_time}
\alias{check_id_and_time}
\title{Check ID and Time Compatibility}
\usage{
check_id_and_time(id, time)
}
\arguments{
\item{id}{An optional vector of unique identifiers corresponding to the time
vector. These can be used to subset the time vector.}

\item{time}{An optional vector of integers of time points observed for each
subject in the sample.}
}
\description{
Check ID and Time Compatibility
}
\keyword{internal}
