% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Combine_gametes.R
\name{Combine_gametes}
\alias{Combine_gametes}
\title{Combine_gametes}
\usage{
Combine_gametes(cohort, father_path, mother_path, ref_path, out_path)
}
\arguments{
\item{cohort}{Your cohort name.}

\item{father_path}{The path of your father gamete file}

\item{mother_path}{The path of your mother gamete file}

\item{ref_path}{The path of your reference file which contains information about A1, A2, SNP and BP}

\item{out_path}{The path of output file}
}
\description{
This function is for combine father gamete and mother gamete
}
\keyword{combine}
