% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaplot.R
\name{densplot.folded}
\alias{densplot.folded}
\title{Densplot Method for Folded}
\usage{
\method{densplot}{folded}(x, ...)
}
\arguments{
\item{x}{folded}

\item{...}{passed to \code{\link{boxplot.data.frame}}}
}
\description{
Densplot method for folded. Converts to data.frame with defined column attributes and calls data.frame method.
}
\examples{
data(eventsf)
library(metaplot)
densplot(eventsf, DV)
}
\seealso{
Other univariate plots: \code{\link{metaplot.folded}}
}
