% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{distill.folded}
\alias{distill.folded}
\title{Distill a Variable of Folded}
\usage{
\method{distill}{folded}(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\description{
Distills a variable of folded.
}
\keyword{internal}
