% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{distill}
\alias{distill}
\title{Distill Something}
\usage{
distill(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\description{
Distills something.
}
\seealso{
\code{\link{fold.data.frame}} \code{\link{distill.folded}} \code{\link{distill.data.frame}}
}
\keyword{internal}
