% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/origami_SuperLearner_helpers.R
\name{drop_zero_learners}
\alias{drop_zero_learners}
\title{drop_zero_learners}
\usage{
drop_zero_learners(osl_fit)
}
\arguments{
\item{osl_fit}{fit generated using orgami_SuperLearner.}

\item{osl_fit}{orgami_SuperLearner without the zero weight learners.}
}
\description{
Drops library learners that were assigned 0 weight, speeding up future predictions.
}

