% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est.R
\name{est}
\alias{est}
\title{est}
\usage{
est(cohort, estfile, chr, snp_path, outpath)
}
\arguments{
\item{cohort}{Your cohort name.}

\item{estfile}{The path of your est file}

\item{chr}{Chromosome number}

\item{snp_path}{The path of your snplist file}

\item{out_path}{The path of output file}
}
\description{
This function is for making est files
}
\keyword{est}
