% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventsf.R
\docType{data}
\name{eventsf}
\alias{eventsf}
\title{Folded Analysis Dataset of Selected Pharmacological Events}
\format{A data frame with 47 rows and 5 variables:
\describe{
  \item{VARIABLE}{name of data item}
  \item{META}{name of metadata item, if any}
  \item{VALUE}{value of item}
  \item{USUBJID}{universal subject identifier, where applicable}
  \item{TIME}{time (h), where applicable}
}}
\source{
derived from 'events'
}
\usage{
eventsf
}
\description{
A data subset of 'events' in folded format
}
\keyword{datasets}
