% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{filter.folded}
\alias{filter.folded}
\title{Filter Folded}
\usage{
\method{filter}{folded}(.data, ...)
}
\arguments{
\item{.data}{passed to next method}

\item{...}{passed to next method}
}
\value{
folded
}
\description{
Filters folded.
}
\keyword{internal}
