% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unfolded.R
\name{filter.unfolded}
\alias{filter.unfolded}
\title{Filter Unfolded}
\usage{
\method{filter}{unfolded}(.data, ...)
}
\arguments{
\item{.data}{passed to next method}

\item{...}{passed to next method}
}
\value{
unfolded
}
\description{
Filters unfolded.
}
\keyword{internal}
