% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{filter_.folded}
\alias{filter_.folded}
\title{Filter Folded}
\usage{
\method{filter_}{folded}(.data, ..., .dots)
}
\arguments{
\item{.data}{passed to next method}

\item{...}{passed to next method}

\item{.dots}{passed to next method}
}
\value{
folded
}
\description{
Filters folded.
}
\keyword{internal}
