% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{fold}
\alias{fold}
\title{Fold an Object}
\usage{
fold(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\description{
Folds an object. Methods are supplied for \code{data.frame} and \code{grouped_df}.
}
\seealso{
\code{\link{fold.data.frame}} 
\code{\link{fold.grouped_df}}
}
\keyword{internal}
