% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{fold_.data.frame}
\alias{fold_.data.frame}
\title{Fold a Data.frame, Standard Evaluation}
\usage{
\method{fold_}{data.frame}(x, ..., args, group_by = groups(x),
  meta = obj_attr(x), simplify = TRUE, sort = TRUE)
}
\arguments{
\item{x}{data.frame}

\item{...}{ignored arguments}

\item{args}{list of formulas representing grouping columns}

\item{group_by}{a vector of column names serving as key: included in result but not stacked}

\item{meta}{a list of formulas in the form object ~ attribute. Pass something with length 0 to suppress guessing.}

\item{simplify}{set to NA any group_by values that do not help distinguish values, and remove resulting duplicate records}

\item{sort}{whether to sort the result}
}
\value{
folded data.frame with columns VARIABLE, META, VALUE and any supplied grouping items.
}
\description{
Folds a data.frame using standard evaluation. See also \code{\link{fold.data.frame}}.
}
\seealso{
\code{\link{obj_attr.data.frame}} \code{\link{fold.data.frame}}
}
\keyword{internal}
