% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{fold_.grouped_df}
\alias{fold_.grouped_df}
\title{Fold a grouped_df, Standard Evaluation}
\usage{
\method{fold_}{grouped_df}(x, groups = unlist(match.fun("groups")(x)),
  meta = obj_attr(x), simplify = TRUE, sort = TRUE, ...)
}
\arguments{
\item{x}{data.frame}

\item{groups}{a vector of column names whose interaction makes records unique. Included in result but not stacked. See also \code{\link{fold.grouped_df}}.  Alternatively, pre-specify as a groups attribute.}

\item{meta}{a list of formulas in the form object ~ attribute. Pass something with length 0 to suppress guessing.}

\item{simplify}{set to NA any groups values that do not help distinguish values, and remove resulting duplicate records}

\item{sort}{whether to sort the result}

\item{...}{passed arguments}
}
\value{
folded data.frame with columns VARIABLE, META, VALUE and any supplied grouping items.
}
\description{
Folds a grouped data.frame using standard evaluation. Reclassifies as data.frame and passes groups explicitly.
}
\seealso{
\code{\link{obj_attr.data.frame}} \code{\link{fold_.data.frame}} \code{\link{fold_}} \code{\link{fold.grouped_df}}
}
