% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{folds2foldvec}
\alias{folds2foldvec}
\title{Build a Fold Vector from a Fold Object}
\usage{
folds2foldvec(folds)
}
\arguments{
\item{folds}{A \code{fold} object as produced by \code{make_folds}, from
which a \code{numeric} vector of the validation set fold IDs are returned.}
}
\description{
For V-fold type cross-validation. This takes a fold object and returns a fold
vector (validation set IDs) for use with other tools like \code{cv.glmnet}.
}
\seealso{
Other fold generation functions: \code{\link{fold_from_foldvec}},
  \code{\link{fold_funs}}, \code{\link{make_folds}},
  \code{\link{make_repeated_folds}}
}
\concept{fold generation functions}
