% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fold_funs.R
\name{folds_bootstrap}
\alias{folds_bootstrap}
\title{Bootstrap cross-validation}
\usage{
folds_bootstrap(n, V = 1000)
}
\arguments{
\item{n}{integer; number of observations}

\item{V}{integer; number of folds to generate.}
}
\description{
Bootstrap cross-validation.
}
\seealso{
Other fold_funs: \code{\link{folds_loo}},
  \code{\link{folds_montecarlo}},
  \code{\link{folds_resubstitution}},
  \code{\link{folds_rolling_origin}},
  \code{\link{folds_rolling_window}},
  \code{\link{folds_vfold}}, \code{\link{make_folds}},
  \code{\link{make_repeated_folds}}
}
