% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fold_funs.R
\name{folds_rolling_window}
\alias{folds_rolling_window}
\title{Rolling window cross-validation}
\usage{
folds_rolling_window(n, window_size, validation_size)
}
\arguments{
\item{n}{integer; number of observations.}

\item{window_size}{integer; number of observations in the training samples.}

\item{validation_size}{integer; number of points in the validation samples. Should be equal to the largest forecast horizon}
}
\description{
Rolling window cross-validation.
}
\seealso{
Other fold_funs: \code{\link{folds_bootstrap}},
  \code{\link{folds_loo}}, \code{\link{folds_montecarlo}},
  \code{\link{folds_resubstitution}},
  \code{\link{folds_rolling_origin}},
  \code{\link{folds_vfold}}, \code{\link{make_folds}}
}
