% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fold_funs.R
\name{folds_vfold}
\alias{folds_vfold}
\title{V-Fold Cross-Validation}
\usage{
folds_vfold(n, V = 10)
}
\arguments{
\item{n}{integer - number of observations.}

\item{V}{integer - number of folds.}
}
\value{
A list of Folds.
}
\description{
Implementation of v-fold cross-validation.
}
\seealso{
Other fold_funs: \code{\link{fold_from_foldvec}},
  \code{\link{folds_bootstrap}}, \code{\link{folds_loo}},
  \code{\link{folds_montecarlo}},
  \code{\link{folds_resubstitution}},
  \code{\link{folds_rolling_origin}},
  \code{\link{folds_rolling_window}},
  \code{\link{make_folds}},
  \code{\link{make_repeated_folds}}
}
