% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_ped.R
\name{gen_ped}
\alias{gen_ped}
\title{gen_ped}
\usage{
gen_ped(familyID, output_path, num_sibling)
}
\arguments{
\item{familyID}{Your cohort name.}

\item{num_sibling}{The number of pseudo sibling you gamete}

\item{out_path}{The path of output file}
}
\description{
This function is for generate ped files for pseudo babies in the format familyID familyID_i
}
\keyword{gen}
