% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{group_by.folded}
\alias{group_by.folded}
\title{Group Folded}
\usage{
\method{group_by}{folded}(.data, ..., add = FALSE)
}
\arguments{
\item{.data}{passed to next method}

\item{...}{passed to next method}

\item{add}{passed to next method}
}
\value{
folded
}
\description{
Groups folded.
}
\keyword{internal}
