% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unfolded.R
\name{group_by.unfolded}
\alias{group_by.unfolded}
\title{Group Unfolded}
\usage{
\method{group_by}{unfolded}(.data, ..., add = FALSE)
}
\arguments{
\item{.data}{passed to next method}

\item{...}{passed to next method}

\item{add}{passed to next method}
}
\value{
unfolded
}
\description{
Groups unfolded. Calls next method, removes \code{groups} attribute and 'unfolded' class.
}
\keyword{internal}
