% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{group_by_.folded}
\alias{group_by_.folded}
\title{Group Folded}
\usage{
\method{group_by_}{folded}(.data, ..., .dots)
}
\arguments{
\item{.data}{passed to next method}

\item{...}{passed to next method}

\item{.dots}{passed to next method}
}
\value{
folded
}
\description{
Groups folded.
}
\keyword{internal}
