% Generated by roxygen2 (4.0.1): do not edit by hand
\name{guess_combiner}
\alias{guess_combiner}
\title{guess_combiner}
\usage{
guess_combiner(result)
}
\arguments{
\item{result}{a single result.}
}
\value{
A function to combine a list of such results.
}
\description{
Maps data types into standard combiners that should be sensible.
}

