% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_folds.R
\name{id_folds_to_folds}
\alias{id_folds_to_folds}
\title{Convert ID Folds to Observation Folds}
\usage{
id_folds_to_folds(idfolds, cluster_ids)
}
\arguments{
\item{idfolds}{folds that subset ids}

\item{cluster_ids}{a vector of cluster ids indicating which observations are in which clusters}
}
\description{
This function convertsf olds that subset ids to folds that subset observations
}
