% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{interpret.folded}
\alias{interpret.folded}
\title{Interpret Data in Folded Format}
\usage{
\method{interpret}{folded}(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\value{
folded data.frame
}
\description{
Interprets data in folded format.  Specifically, substitutes decodes for codes
when presenting encoded variables.
}
