% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{left_join.folded}
\alias{left_join.folded}
\title{Left-join Folded}
\usage{
\method{left_join}{folded}(.data, ..., .dots)
}
\arguments{
\item{.data}{passed to next method}

\item{...}{passed to next method}

\item{.dots}{passed to next method}
}
\value{
folded
}
\description{
Left-joins folded.
}
\keyword{internal}
