% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unfolded.R
\name{left_join.unfolded}
\alias{left_join.unfolded}
\title{Left-join Unfolded}
\usage{
\method{left_join}{unfolded}(.data, ..., .dots)
}
\arguments{
\item{.data}{passed to next method}

\item{...}{passed to next method}

\item{.dots}{passed to next method}
}
\value{
unfolded
}
\description{
Left-joins unfolded.
}
\keyword{internal}
