% Generated by roxygen2 (4.0.1): do not edit by hand
\name{make_folds}
\alias{make_folds}
\title{Make list of folds for cross-validation}
\usage{
make_folds(n = NULL, fold_fun = NULL, cluster_ids = NULL,
  strata_ids = NULL, ...)
}
\arguments{
\item{n}{either an integer indicating the number of observations to cross-validate over, or an object from which to guess the number of observations. Can also be computed from strata_ids or cluster_ids.}

\item{fold_fun}{either a character vector or a function indicating the cross-validation scheme to use.}

\item{cluster_ids}{a vector of cluster ids. Clusters are treated as a unit - they are all placed in either the training or validation set.}

\item{strata_ids}{a vector of strata ids. Strata are balanced - insofar as possible the distribution in the sample should be the same as the distribution in the training and validation sets.}

\item{...}{other arguments to \code{fold_fun}.}
}
\value{
A list of Folds.
}
\description{
Generates a list of folds for a variety of cross-validation schemes.
}
\seealso{
\code{link{Fold}}

Other fold_funs: \code{\link{folds_bootstrap}};
  \code{\link{folds_loo}}; \code{\link{folds_montecarlo}};
  \code{\link{folds_vfold}}
}

