% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_map.R
\name{make_map}
\alias{make_map}
\title{make_map}
\usage{
make_map(chr, ref_file, rs_file, output_path)
}
\arguments{
\item{chr}{chromosome number}

\item{ref_file}{Ref file of your vcf data}

\item{rs_file}{Reference file contains SNP and BP information}

\item{output_path}{output file path}
}
\description{
This function is to generate map files to map BP and rsid together
}
\keyword{map}
