% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meiosis_local.R
\name{meiosis_local}
\alias{meiosis_local}
\title{meiosis_local}
\usage{
meiosis_local(ind_data, cohort, est_ori_file, est_new_file, num_sibling,
  output_path, interference_number = 3)
}
\arguments{
\item{ind_data}{Your cohort name.}

\item{cohort}{The number of pseudo sibling you gamete}

\item{est_ori_file}{The path of original est file}

\item{est_new_file}{The path of processed est file}

\item{num_sibling}{Number of sibling you gamete}
}
\description{
This function is for generate ped files for pseudo babies in the format familyID familyID_i
}
\keyword{local}
