% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaplot.R
\name{metaplot.folded}
\alias{metaplot.folded}
\title{Create Metaplot from Folded}
\usage{
\method{metaplot}{folded}(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\description{
Creates a plot from folded.  Packs metadata into attributes and calls next method.
}
\seealso{
Other univariate plots: \code{\link{dens.folded}}

Other bivariate plots: \code{\link{boxplot.folded}},
  \code{\link{scatter.folded}}

Other multivariate plots: \code{\link{corsplom.folded}}
}
