% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/multinomial_SL.R
\name{multinomial_SuperLearner}
\alias{multinomial_SuperLearner}
\title{multinomial_SuperLearner}
\usage{
multinomial_SuperLearner(Y, X, SL.library = c("mnSL.randomForest",
  "mnSL.glmnet", "mnSL.multinom", "mnSL.mean"), ...)
}
\description{
Convenience function to specify mulitnomial wrappers, family, and combination method.
       See \code{\link{predict.origami_SuperLearner}} for details of other arguments
}
\seealso{
\code{\link{predict.origami_SuperLearner}}
}

