% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folded.R
\name{mutate.folded}
\alias{mutate.folded}
\title{Mutate Folded}
\usage{
\method{mutate}{folded}(.data, ...)
}
\arguments{
\item{.data}{passed to next method}

\item{...}{passed to next method}

\item{.dots}{passed to next method}
}
\value{
folded
}
\description{
Mutates folded.
}
\keyword{internal}
