% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unfolded.R
\name{mutate.unfolded}
\alias{mutate.unfolded}
\title{Mutate Unfolded}
\usage{
\method{mutate}{unfolded}(.data, ...)
}
\arguments{
\item{.data}{passed to next method}

\item{...}{passed to next method}

\item{.dots}{passed to next method}
}
\value{
unfolded
}
\description{
Mutates unfolded.
}
\keyword{internal}
