% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/origami-package.R
\docType{package}
\name{origami-package}
\alias{origami-package}
\alias{origami}
\title{origami - Cross-validation Framework}
\description{
origami - Cross-validation Framework
}
\details{
\tabular{ll}{
Package: \tab origami\cr
Type: \tab Package\cr
Version: \tab 0.5\cr
Date: \tab 2014-08-17\cr
License: \tab GPL-3\cr
LazyLoad: \tab yes\cr

}

origami provides a general framework for the application of
cross-validation schemes to particular functions. By allowing arbitrary
lists of results, origami accommodates a range of cross-validation
applications. Some applications are implemented as examples.
}
\examples{
data(mtcars)

# resubstitution MSE
r <- lm(mpg ~ ., data = mtcars)
mean(resid(r)^2)

# function to calculate cross-validated squared error
cvlm <- function(fold) {
    train_data <- training(mtcars)
    valid_data <- validation(mtcars)

    r <- lm(mpg ~ ., data = train_data)
    preds <- predict(r, newdata = valid_data)
    list(coef = data.frame(t(coef(r))), SE = ((preds - valid_data$mpg)^2))

}

# replicate the resubstitution estimate
resub <- make_folds(mtcars, fold_fun = "resubstitution")[[1]]
resub_results <- cvlm(resub)
mean(resub_results$SE)

# cross-validated estimate
folds <- make_folds(mtcars)
results <- cross_validate(cvlm, folds)
mean(results$SE)
}
\seealso{
\code{\link{cross_validate}}, \code{\link{origami_SuperLearner}}
}
\author{
Jeremy Coyle \email{jeremyrcoyle@gmail.com}
}
\keyword{package}
