% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/origami_SuperLearner.R
\name{predict.origami_SuperLearner}
\alias{predict.origami_SuperLearner}
\title{predict.origami_SuperLearner}
\usage{
\method{predict}{origami_SuperLearner}(object, newdata = "cv-original", ...)
}
\arguments{
\item{object}{origami_SuperLearner fit.}

\item{newdata}{matrix or data.frame of new covariate values. If newdata='cv-original' (default), it will return Z (the split-specific library predictions), and the Super Learner applied to Z (the split-specific Super Learner Predictions)}

\item{...}{other arguments to the learners.}
}
\value{
A list with two elements: \code{library_pred} are the predictions from the library functions, and \code{pred} is the prediction from the SuperLearner.
}
\description{
prediction function for origami_SuperLearner. Note, while this is a working SuperLearner implementation, it is intended more as an example 
than production code. As such, it is subject to change in the future.
}
\seealso{
\code{\link{origami_SuperLearner}}
}
