% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{print.folded}
\alias{print.folded}
\title{Print Folded}
\usage{
\method{print}{folded}(x, limit = 8, ...)
}
\arguments{
\item{x}{folded}

\item{limit}{number of characters to allow without intervention}

\item{...}{passed arguments}
}
\value{
character
}
\description{
Prints folded. Specifically, shortens display of encoded items that are above limit.
}
\keyword{internal}
