% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaplot.R
\name{scatter.folded}
\alias{scatter.folded}
\title{Scatterplot for Folded}
\usage{
\method{scatter}{folded}(x, ...)
}
\arguments{
\item{x}{folded}

\item{...}{unquoted names of variables to plot, or other named arguments}
}
\description{
Scatterplot for class 'folded'.
}
\seealso{
Other bivariate plots: \code{\link{boxplot.folded}},
  \code{\link{metaplot.folded}}
}
