% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{select.folded}
\alias{select.folded}
\title{Select Folded}
\usage{
\method{select}{folded}(.data, ...)
}
\arguments{
\item{.data}{passed to next method}

\item{...}{passed to next method}
}
\value{
folded
}
\description{
Selects folded.
}
\keyword{internal}
