% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unfolded.R
\name{select.unfolded}
\alias{select.unfolded}
\title{Select Unfolded}
\usage{
\method{select}{unfolded}(.data, ...)
}
\arguments{
\item{.data}{passed to next method}

\item{...}{passed to next method}
}
\value{
unfolded
}
\description{
Selects unfolded.
}
\keyword{internal}
