% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{select_.folded}
\alias{select_.folded}
\title{Select Folded}
\usage{
\method{select_}{folded}(.data, ..., .dots)
}
\arguments{
\item{.data}{passed to next method}

\item{...}{passed to next method}

\item{.dots}{passed to next method}
}
\value{
folded
}
\description{
Selects folded.
}
\keyword{internal}
