% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folded.R
\name{semi_join.folded}
\alias{semi_join.folded}
\title{Semi-join Folded}
\usage{
\method{semi_join}{folded}(.data, ..., .dots)
}
\arguments{
\item{.data}{passed to next method}

\item{...}{passed to next method}

\item{.dots}{passed to next method}
}
\value{
folded
}
\description{
Semi-joins folded.
}
\keyword{internal}
