% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unfolded.R
\name{semi_join.unfolded}
\alias{semi_join.unfolded}
\title{Semi-join Unfolded}
\usage{
\method{semi_join}{unfolded}(.data, ..., .dots)
}
\arguments{
\item{.data}{passed to next method}

\item{...}{passed to next method}

\item{.dots}{passed to next method}
}
\value{
unfolded
}
\description{
Semi-joins unfolded.
}
\keyword{internal}
