% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{sort.folded}
\alias{sort.folded}
\title{Sort Folded}
\usage{
\method{sort}{folded}(x, decreasing = FALSE, ...)
}
\arguments{
\item{x}{folded}

\item{decreasing}{logical}

\item{...}{ignored arguments}
}
\value{
folded
}
\description{
Sorts folded.
}
