% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folded.R
\name{transmutate.folded}
\alias{transmutate.folded}
\title{Transmutate Folded}
\usage{
transmutate.folded(.data, ...)
}
\arguments{
\item{.data}{passed to next method}

\item{...}{passed to next method}

\item{.dots}{passed to next method}
}
\value{
folded
}
\description{
Transmutates folded.
}
\keyword{internal}
