% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unfolded.R
\name{transmutate.unfolded}
\alias{transmutate.unfolded}
\title{Transmutate Unfolded}
\usage{
transmutate.unfolded(.data, ...)
}
\arguments{
\item{.data}{passed to next method}

\item{...}{passed to next method}

\item{.dots}{passed to next method}
}
\value{
unfolded
}
\description{
Transmutates unfolded.
}
\keyword{internal}
