% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{unfold.folded}
\alias{unfold.folded}
\title{Unfold a Folded Format Object}
\usage{
\method{unfold}{folded}(x, var = unique(x$VARIABLE[is.na(x$META)]), ...)
}
\arguments{
\item{x}{object}

\item{var}{variables to unfold, given as character vector}

\item{...}{passed arguments}
}
\description{
Unfolds a folded data.frame.  Ideally it should be losslessly reversible.
}
