% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{unfold_.folded}
\alias{unfold_.folded}
\title{Unfold a Folded Data.frame, Standard Evaluation}
\usage{
\method{unfold_}{folded}(x, ..., var)
}
\arguments{
\item{x}{folded data.frame}

\item{...}{ignored arguments}

\item{var}{character: variables to unfold}
}
\value{
data.frame
}
\description{
Unfolds a folded data.frame, or part thereof, using standard evaluation.
}
\details{
By default, the entire data.frame is unfolded, possibly giving back something originally passed to fold().  If \code{var} is specified, only selected items are unfolded.  Values stored as encodings are converted to factor.
}
\seealso{
fold.data.frame
}
\keyword{internal}
