% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{unfold_}
\alias{unfold_}
\title{Unfold an Object, Standard Evaluation}
\usage{
unfold_(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\description{
Unfolds an object using standard evaluation.
}
\seealso{
\code{\link{unfold_.folded}} \code{\link{unfold}}
}
\keyword{internal}
