% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folded.R
\name{ungroup.folded}
\alias{ungroup.folded}
\title{Ungroup Folded}
\usage{
\method{ungroup}{folded}(.data, ..., .dots)
}
\arguments{
\item{.data}{passed to next method}

\item{...}{passed to next method}

\item{.dots}{passed to next method}
}
\value{
folded
}
\description{
Ungroups folded.
}
\keyword{internal}
