% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unfolded.R
\name{ungroup.unfolded}
\alias{ungroup.unfolded}
\title{Ungroup Unfolded}
\usage{
\method{ungroup}{unfolded}(.data, ..., .dots)
}
\arguments{
\item{.data}{passed to next method}

\item{...}{passed to next method}

\item{.dots}{passed to next method}
}
\value{
unfolded
}
\description{
Ungroups unfolded. Calls next method, removes \code{groups} attribute and 'unfolded' class.
}
\keyword{internal}
