% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaplot.R
\name{unpack.folded}
\alias{unpack.folded}
\title{Unpack a Folded Data Frame}
\usage{
\method{unpack}{folded}(x, tolower = TRUE, ...)
}
\arguments{
\item{x}{folded}

\item{tolower}{whether to coerce attribute names to lower case}

\item{...}{other arguments}
}
\value{
data.frame
}
\description{
Convert folded data.frame to unpacked format with scalar metadata as row entries.
}
\examples{
data(eventsf)
library(metaplot)
head(unpack(eventsf))

}
\seealso{
\code{\link{fold.data.frame}} \code{\link{pack.folded}}
}
