% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{wrap_in_try}
\alias{wrap_in_try}
\title{Wrap a Function in a Try Statement}
\usage{
wrap_in_try(fun, ...)
}
\arguments{
\item{fun}{A \code{function} to be wrapped in a \code{try} statement.}

\item{...}{Additional arguments passed to the previous argument \code{fun}.}
}
\description{
Function factory that generates versions of functions wrapped in \code{try}.
}
