% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{as.folded.character}
\alias{as.folded.character}
\title{Coerce to Folded from Character}
\usage{
\method{as.folded}{character}(x, ...)
}
\arguments{
\item{x}{length-one character}

\item{...}{passed arguments}
}
\value{
folded dat.frame
}
\description{
Coerces to folded by treating character as a filepath.
}
\examples{
library(magrittr)
library(csv)
data(eventsf)
file <- tempfile()
eventsf \%>\% as.csv(file)
as.folded(file)
}
\seealso{
\code{\link{as.csv.folded}} \code{\link{as.folded}}
}
