% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{as.folded.folded}
\alias{as.folded.folded}
\title{Coerce to Folded Format from Folded Format}
\usage{
\method{as.folded}{folded}(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\value{
folded
}
\description{
Coerces to folded format from folded.  A non-operation.
}
\keyword{internal}
