% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{distill.folded}
\alias{distill.folded}
\title{Distill Folded}
\usage{
\method{distill}{folded}(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\description{
Distills folded. Indirectly calls the data.frame method.
}
\seealso{
\code{\link{distill}}
}
\keyword{internal}
