% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{fold.data.frame}
\alias{fold.data.frame}
\title{Fold a Data Frame}
\usage{
\method{fold}{data.frame}(x, ..., meta = obj_attr(x), simplify = TRUE,
  sort = TRUE)
}
\arguments{
\item{x}{data.frame}

\item{...}{unquoted names of grouping columns}

\item{meta}{a list of formulas in the form object ~ attribute. Pass something with length 0 to suppress guessing.}

\item{simplify}{set to NA any groups values that do not help distinguish values, and remove resulting duplicate records}

\item{sort}{whether to sort the result}
}
\value{
folded data.frame with columns VARIABLE, META, VALUE and any supplied grouping items.
}
\description{
Folds a data.frame.  Stacks columns, while isolating metadata and capturing keys.
}
\seealso{
\code{\link{obj_attr.data.frame}},  \code{\link{fold_.data.frame}}
}
