% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{fold.grouped_df}
\alias{fold.grouped_df}
\title{Fold grouped_df}
\usage{
\method{fold}{grouped_df}(x, ...)
}
\arguments{
\item{x}{grouped_df}

\item{...}{passed arguments}
}
\value{
folded
}
\description{
Folds grouped_df, unclassing and passing groups explcitly as groups(x), over-ridden by anonymous arguments.
}
\seealso{
\code{\link{fold.data.frame}}
}
