% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{fold}
\alias{fold}
\title{Fold an Object}
\usage{
fold(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\description{
Folds an object. A method is supplied for \code{data.frame}. The actual work is done by \code{\link{fold_.data.frame}}.
}
\seealso{
\code{\link{fold.data.frame}}
}
\keyword{internal}
