% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Fold.R
\docType{class}
\name{Fold}
\alias{Fold}
\alias{make_fold}
\title{Fold S4 class}
\usage{
make_fold(v, training_set, validation_set)
}
\arguments{
\item{v}{integer index of fold in larger scheme.}

\item{training_set}{integer vector of indexes corresponding to the training set.}

\item{validation_set}{integer vector of indexes corresponding to the validation set.}
}
\value{
A Fold.
}
\description{
class representing a fold in a cross-validation scheme
}
\seealso{
\code{\link{fold_helpers}}
}

