% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{fold_}
\alias{fold_}
\title{Fold an Object, Standard Evaluation}
\usage{
fold_(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\description{
Folds an object using standard evaluation.
}
\seealso{
\code{\link{fold}} \code{\link{fold_.data.frame}} \code{\link{fold_.grouped_df}}
}
\keyword{internal}
