% Generated by roxygen2 (4.0.1): do not edit by hand
\name{folds_bootstrap}
\alias{folds_bootstrap}
\title{Bootstrap cross-validation}
\usage{
folds_bootstrap(n, V = 1000)
}
\arguments{
\item{n}{integer; number of observations.either an integer indicating the number of observations to cross-validate over, or an object from which to guess the number of observations. Can also be computed from strata_ids or cluster_ids.}

\item{V}{integer; number of folds to generate.}
}
\description{
Bootstrap cross-validation.
}
\seealso{
Other fold_funs: \code{\link{folds_loo}};
  \code{\link{folds_montecarlo}};
  \code{\link{folds_vfold}}; \code{\link{make_folds}}
}

