% Generated by roxygen2 (4.0.1): do not edit by hand
\name{folds_loo}
\alias{folds_loo}
\title{Leave-one-out cross-validation}
\usage{
folds_loo(n)
}
\arguments{
\item{n}{integer; number of observations.}
}
\value{
A list of Folds
}
\description{
Leave-one-out cross-validation.
}
\seealso{
Other fold_funs: \code{\link{folds_bootstrap}};
  \code{\link{folds_montecarlo}};
  \code{\link{folds_vfold}}; \code{\link{make_folds}}
}

