% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fold_funs.R
\name{folds_montecarlo}
\alias{folds_montecarlo}
\title{Monte Carlo cross-validation}
\usage{
folds_montecarlo(n, V = 1000, pvalidation = 0.2)
}
\arguments{
\item{n}{integer; number of observations.}

\item{V}{integer; number of folds to generate.}

\item{pvalidation}{proportion of observation to be in validation fold.}
}
\description{
Monte Carlo cross-validation.
}
\seealso{
Other fold_funs: \code{\link{folds_bootstrap}};
  \code{\link{folds_loo}};
  \code{\link{folds_resubstitution}};
  \code{\link{folds_rolling_origin}};
  \code{\link{folds_rolling_window}};
  \code{\link{folds_vfold}}; \code{\link{make_folds}}
}

