% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fold_funs.R
\name{folds_rolling_origin}
\alias{folds_rolling_origin}
\title{Rolling Origin Cross-Validation}
\usage{
folds_rolling_origin(n, first_window, validation_size, gap = 0, batch = 1)
}
\arguments{
\item{n}{(integer) - number of observations.}

\item{first_window}{(integer) - number of observations in the first training
sample.}

\item{validation_size}{(integer) - number of points in the validation
samples; should be equal to the largest forecast horizon.}

\item{gap}{(integer) - number of points not included in the training or validation 
samples; Default is 0.}

\item{batch}{(integer) - Increases the number of time-points added to the training
set each CV iteration. Applicable for larger time-series. Default is 1.}
}
\description{
Implementation of rolling origin cross-validation.
}
\seealso{
Other fold_funs: \code{\link{fold_from_foldvec}},
  \code{\link{folds_bootstrap}}, \code{\link{folds_loo}},
  \code{\link{folds_montecarlo}},
  \code{\link{folds_resubstitution}},
  \code{\link{folds_rolling_window}},
  \code{\link{folds_vfold}}, \code{\link{make_folds}},
  \code{\link{make_repeated_folds}}
}
