% Generated by roxygen2 (4.0.1): do not edit by hand
\name{folds_vfold}
\alias{folds_vfold}
\title{V-fold cross-validation}
\usage{
folds_vfold(n, V = 10)
}
\arguments{
\item{n}{integer; number of observations.}

\item{V}{integer; number of folds.}
}
\value{
A list of Folds.
}
\description{
V-fold cross-validation.
}
\seealso{
Other fold_funs: \code{\link{folds_bootstrap}};
  \code{\link{folds_loo}}; \code{\link{folds_montecarlo}};
  \code{\link{folds_rolling_origin}};
  \code{\link{folds_rolling_window}};
  \code{\link{make_folds}}
}

