% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_results.R
\name{guess_combiner}
\alias{guess_combiner}
\title{Flexible Guessing and Mapping for Combining Data Types}
\usage{
guess_combiner(result)
}
\arguments{
\item{result}{- a single result; flexibly accepts several object classes.}
}
\value{
A function to combine a list of such results.
}
\description{
Maps data types into standard combiners that should be sensible.
}
