% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fold.R
\name{make_fold}
\alias{make_fold}
\title{Fold}
\usage{
make_fold(v, training_set, validation_set)
}
\arguments{
\item{v}{An integer index of folds in the larger scheme.}

\item{training_set}{An integer vector of indexes corresponding to the
training set.}

\item{validation_set}{An integer vector of indexes corresponding to the
validation set.}
}
\value{
A list containing these elements.
}
\description{
Functions to make a fold. Current representation is a simple \code{list}.
}
\seealso{
\code{\link{fold_helpers}}
}
