% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fold_funs.R
\name{make_repeated_folds}
\alias{make_repeated_folds}
\title{Repeated Cross-Validation}
\usage{
make_repeated_folds(repeats, ...)
}
\arguments{
\item{repeats}{integer; number of repeats}

\item{...}{arguments passed to \link{make_folds}}
}
\description{
Implementation of repeated window cross-validation: generates fold objects
for repeated cross-validation by making repeated calls to \link{make_folds}
and concatenating the results.
}
\seealso{
Other fold_funs: \code{\link{fold_from_foldvec}},
  \code{\link{folds_bootstrap}}, \code{\link{folds_loo}},
  \code{\link{folds_montecarlo}},
  \code{\link{folds_resubstitution}},
  \code{\link{folds_rolling_origin}},
  \code{\link{folds_rolling_window}},
  \code{\link{folds_vfold}}, \code{\link{make_folds}}
}
