% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{sort.folded}
\alias{sort.folded}
\title{Sort Folded}
\usage{
\method{sort}{folded}(x, decreasing = FALSE, ...)
}
\arguments{
\item{x}{folded}

\item{decreasing}{logical}

\item{...}{ignored arguments}
}
\value{
folded
}
\description{
Sorts folded, using all non-VALUE columns, starting from the left.  If decreasing is TRUE, the resulting row order is reversed.
}
\examples{
library(magrittr)
data(eventsf)
eventsf \%>\% sort
}
\seealso{
\code{\link{fold.data.frame}}
}
