% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{unfold.folded}
\alias{unfold.folded}
\title{Unfold a Folded Data.frame}
\usage{
\method{unfold}{folded}(x, ...)
}
\arguments{
\item{x}{folded data.frame}

\item{...}{variables to unfold, given as unquoted anonymous names}
}
\value{
data.frame with a groups attribute (character)
}
\description{
Unfolds a folded data.frame, or part thereof.
}
\details{
By default, the entire data.frame is unfolded, possibly giving back something originally passed to fold().  If \dots is specified, only selected items (given as anonymous unquoted arguments) are unfolded.  Values stored as encodings are converted to factor. The result has a groups attribute: a character vector of column names in the result whose interaction makes rows unique.
}
\seealso{
\code{\link{fold_.data.frame}}
}
