% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{ungroup.folded}
\alias{ungroup.folded}
\title{Ungroup Folded}
\usage{
\method{ungroup}{folded}(.data, ..., .dots)
}
\arguments{
\item{.data}{A tbl. All main verbs are S3 generics and provide methods
for \code{\link{tbl_df}}, \code{\link[dtplyr]{tbl_dt}} and \code{\link{tbl_sql}}.}

\item{...}{Logical predicates. Multiple conditions are combined with \code{&}.}

\item{.dots}{Used to work around non-standard evaluation. See
\code{vignette("nse")} for details.}
}
\value{
folded
}
\description{
Ungroups folded.
}
